# Conflicts:
#	src/main/java/ru/skillbox/socialnetwork/api/dto/PostParameters.java
#	src/main/java/ru/skillbox/socialnetwork/model/Dialog.java
#	src/main/java/ru/skillbox/socialnetwork/model/Message.java

+++ src/main/java/ru/skillbox/socialnetwork/service/PlatformService.java
@@ -13,14 +13,14 @@ public class PlatformService {
     @Autowired
     PlatformDAO platformDAO;
 
-    public AbstractResponse languages(String language, int offset,  int itemPerPage){
+    public AbstractResponse languages(String language, Long offset,  Long itemPerPage){
 
         AbstractResponse response = new ListResponseApi("string",
-                System.currentTimeMillis(),
-                new PlatfomLanguageApi(1, "Русский"),
-                0,
-                0,
-                0);
+                    System.currentTimeMillis(),
+                    new PlatfomLanguageApi(1, "Русский"),
+                    0,
+                    0,
+                    0);
 
         return response;
     }
@@ -28,11 +28,11 @@ public class PlatformService {
     public AbstractResponse countries(String country, int offset, int itemPerPage){
 
         AbstractResponse response = new ListResponseApi("string",
-                System.currentTimeMillis(),
-                new PlatfomLanguageApi(1, "Россия"),
-                0,
-                0,
-                0);
+                    System.currentTimeMillis(),
+                    new PlatfomLanguageApi(1, "Россия"),
+                    0,
+                    0,
+                    0);
 
         return response;
     }
